requirements.txt
# --- Frontend ---
streamlit

# --- Backend API ---
fastapi
uvicorn
python-multipart

# --- Image & Data Processing ---
# Pinned to <2.0 due to TFLite compatibility
numpy<2
pillow
opencv-python-headless

# --- Inference Engine (Conditional Installation) ---
# Use lightweight runtime for production (Linux/Docker)
tflite-runtime; platform_system=="Linux"

# Use full TensorFlow for local development (Windows/MacOS)
tensorflow; platform_system!="Linux"

# --- Utilities ---
gdown
requests